LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/api/cors/cors-basic.any-expected.txt

CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Fetch API cannot load http://web-platform.test:8801/fetch/api/resources/top.txt due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/fetch/api/resources/top.txt due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8800/fetch/api/resources/top.txt due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Fetch API cannot load http://www1.web-platform.test:8801/fetch/api/resources/top.txt due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Fetch API cannot load https://www1.web-platform.test:9443/fetch/api/resources/top.txt due to access control checks.

PASS Same domain different port [no-cors mode]
PASS Same domain different port [server forbid CORS]
PASS Same domain different port [cors mode]
PASS Same domain different protocol different port [no-cors mode]
PASS Same domain different protocol different port [server forbid CORS]
PASS Same domain different protocol different port [cors mode]
PASS Cross domain basic usage [no-cors mode]
PASS Cross domain basic usage [server forbid CORS]
PASS Cross domain basic usage [cors mode]
PASS Cross domain different port [no-cors mode]
PASS Cross domain different port [server forbid CORS]
PASS Cross domain different port [cors mode]
PASS Cross domain different protocol [no-cors mode]
PASS Cross domain different protocol [server forbid CORS]
PASS Cross domain different protocol [cors mode]
